+++ private/typeAstToMdAst.js
@@ -192,7 +192,7 @@ module.exports = function typeAstToMdAst(typeJsdocAst, members) {
       break;
     }
     default:
-      throw new Error(`Unknown JSDoc type \`${typeJsdocAst.type}\`.`);
+      throw new TypeError(`Unknown JSDoc type \`${typeJsdocAst.type}\`.`);
   }
 
   return children;